Basically finished. Needs to be tested on an fpga
I made changes in the seven segment decoder file, where I added the outputs for inputs 10-15 in hexidecimal. The mux2to1 file now accepts a 16 bit bus as an input and outputs a 16 bit bus. The top level uses the mux to select between the switches or the bcd numbers. Test benches still need to be made.